requirements.txt
# GCP SQL
sqlalchemy

# Data
numpy
pandas

# Utils
python-dotenv
